+++ packages/api/src/trueppm_api/urls.py
@@ -21,4 +21,5 @@ urlpatterns = [
     path("api/v1/", include("trueppm_api.apps.resources.urls")),
     path("api/v1/", include("trueppm_api.apps.scheduling.urls")),
     path("api/v1/", include("trueppm_api.apps.sync.urls")),
+    path("api/v1/", include("trueppm_api.apps.history.urls")),
 ]